Handle potential overflow

+++ src/util/Message.cpp
@@ -667,7 +667,7 @@ bool Message::testLength(unsigned int length) noexcept {
 }
 
 unsigned int Message::packets(unsigned int bytes) noexcept {
-	return ((size_t) bytes + PAYLOAD_SIZE - 1) / PAYLOAD_SIZE;
+	return ((unsigned long long) bytes + PAYLOAD_SIZE - 1) / PAYLOAD_SIZE;
 }
 
 } /* namespace wanhive */
